+++ packages/strapi-design-system/src/themes/index.ts
@@ -1,4 +1,12 @@
+import { Colors, Shadows } from './colors';
+import { CommonTheme } from './common-theme';
+
 export * from './lightTheme';
 export * from './darkTheme';
 export * from './extendTheme';
 export * from './utils';
+
+export interface StrapiTheme extends CommonTheme {
+  colors: Colors;
+  shadows: Shadows;
+}
